:合并

+++ plugin/guessmusic/struct.go
@@ -168,25 +168,25 @@ type topList struct {
 			} `json:"al"`
 			Dt int `json:"dt"`
 			H  struct {
-				Br   int `json:"br"`
-				Fid  int `json:"fid"`
-				Size int `json:"size"`
-				Vd   int `json:"vd"`
-				Sr   int `json:"sr"`
+				Br   int     `json:"br"`
+				Fid  int     `json:"fid"`
+				Size int     `json:"size"`
+				Vd   float64 `json:"vd"`
+				Sr   int     `json:"sr"`
 			} `json:"h"`
 			M struct {
-				Br   int `json:"br"`
-				Fid  int `json:"fid"`
-				Size int `json:"size"`
-				Vd   int `json:"vd"`
-				Sr   int `json:"sr"`
+				Br   int     `json:"br"`
+				Fid  int     `json:"fid"`
+				Size int     `json:"size"`
+				Vd   float64 `json:"vd"`
+				Sr   int     `json:"sr"`
 			} `json:"m"`
 			L struct {
-				Br   int `json:"br"`
-				Fid  int `json:"fid"`
-				Size int `json:"size"`
-				Vd   int `json:"vd"`
-				Sr   int `json:"sr"`
+				Br   int     `json:"br"`
+				Fid  int     `json:"fid"`
+				Size int     `json:"size"`
+				Vd   float64 `json:"vd"`
+				Sr   int     `json:"sr"`
 			} `json:"l"`
 			Sq                   interface{}   `json:"sq"`
 			Hr                   interface{}   `json:"hr"`
